+++ Mahmoud/Fetching-and-Updating-Data-with-React-Query/react-course-part2-starter-main/src/react-query/constants.ts
@@ -1,0 +1,1 @@
+export const CACHE_KEY_TODOS = ["todos"];
