+++ Mahmoud/Fetching-and-Updating-Data-with-React-Query/react-course-part2-starter-main/src/react-query/hooks/useAddTodo.ts
@@ -1,0 +1,49 @@
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { Todo } from "./useTodos";
+import axios from "axios";
+import { CACHE_KEY_TODOS } from "../constants";
+
+interface AddTOdoContext {
+  previousTodos: Todo[];
+}
+interface AddTOdoContext {
+  previousTodos: Todo[];
+}
+const useAddTodo = (onAdd: () => void) => {
+  const queryClient = useQueryClient();
+  return useMutation<Todo, Error, Todo, AddTOdoContext>({
+    mutationFn: (todo: Todo) =>
+      axios
+        .post<Todo>("https://jsonplaceholder.typicode.com/todos", todo)
+        .then((res) => res.data),
+    onMutate(newTodo: Todo) {
+      const previousTodos =
+        queryClient.getQueryData<Todo[]>(CACHE_KEY_TODOS) || [];
+      queryClient.setQueryData<Todo[]>(CACHE_KEY_TODOS, (todos = []) => [
+        newTodo,
+        ...todos,
+      ]);
+
+      onAdd();
+
+      return { previousTodos };
+    },
+
+    onSuccess: (savedTodo, newTodo) => {
+      queryClient.setQueriesData<Todo[]>(["todo"], (todos) =>
+        todos?.map((todo) => (todo === newTodo ? savedTodo : todo))
+      );
+    },
+
+    onError(error, newTodo, context) {
+      if (!context) return;
+
+      queryClient.setQueriesData<Todo[]>(
+        CACHE_KEY_TODOS,
+        context.previousTodos
+      );
+    },
+  });
+};
+
+export default useAddTodo;
